LayoutTests/fast/forms/ios/keyboard-restoration-after-element-replacement-expected.txt

Test keyboard restoration when input elements are replaced during the same presentation update.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isShowingKeyboard is true
PASS isShowingKeyboard is true
PASS isShowingKeyboard is false
PASS isShowingKeyboard is true
PASS isShowingKeyboard is false
PASS isShowingKeyboard is true
PASS isShowingKeyboard is false
PASS successfullyParsed is true

TEST COMPLETE
